refactor: move TableForm component in dir structure
